+++ apps/send-func/src/func.ts
@@ -1,10 +1,37 @@
 import { app } from "@azure/functions";
+import { loadConfigFromEnvironment } from "io-messages-common/adapters/config";
+import LollipopClient from "io-messages-common/adapters/lollipop/lollipop-client";
+import { createLollipopMiddleware } from "io-messages-common/adapters/lollipop/lollipop-middleware";
+import { handlerWithMiddleware } from "io-messages-common/adapters/middleware";
 
+import { Config, configFromEnvironment } from "./adapters/config.js";
+import { getNotification } from "./adapters/functions/aar-notifications.js";
+import { aarQRCodeCheck } from "./adapters/functions/aar-qrcode-check.js";
 import { healthcheck } from "./adapters/functions/health.js";
+import NotificationClient from "./adapters/send/notification.js";
 import { HealthUseCase } from "./domain/use-cases/health.js";
 
-const main = async (): Promise<void> => {
+const main = async (config: Config): Promise<void> => {
   const healthcheckUseCase = new HealthUseCase([]);
+
+  const notificationClient = new NotificationClient(
+    config.notificationClient.apiKey,
+    config.notificationClient.baseUrl,
+  );
+
+  const uatNotificationClient = new NotificationClient(
+    config.notificationUatClient.apiKey,
+    config.notificationUatClient.baseUrl,
+  );
+
+  const getSendClient = (isTest: boolean) =>
+    isTest ? uatNotificationClient : notificationClient;
+
+  const lollipopClient = new LollipopClient(
+    config.lollipop.apiKey,
+    config.lollipop.baseUrl,
+  );
+  const lollipopMiddleware = createLollipopMiddleware(lollipopClient);
 
   app.http("Health", {
     authLevel: "anonymous",
@@ -12,6 +39,26 @@ const main = async (): Promise<void> => {
     methods: ["GET"],
     route: "health",
   });
+
+  app.http("AARQrCodeCheck", {
+    authLevel: "anonymous",
+    handler: handlerWithMiddleware(
+      lollipopMiddleware,
+      aarQRCodeCheck(getSendClient),
+    ),
+    methods: ["POST"],
+    route: "aar/qr-code-check",
+  });
+
+  app.http("GetAARNotification", {
+    authLevel: "anonymous",
+    handler: handlerWithMiddleware(
+      lollipopMiddleware,
+      getNotification(getSendClient),
+    ),
+    methods: ["GET"],
+    route: "aar/notifications/{iun}",
+  });
 };
 
-await main();
+await loadConfigFromEnvironment(main, configFromEnvironment);